+++ pkg/core/graph/graphtest/graphtest.go
@@ -5,9 +5,12 @@ package graphtest
 
 import (
 	"fmt"
+	"os"
+	"slices"
 	"sync"
 	"testing"
 
+	"github.com/gomlx/go-xla/pkg/installer"
 	"github.com/gomlx/gomlx/backends"
 	"github.com/gomlx/gomlx/backends/xla"
 	"github.com/gomlx/gomlx/pkg/core/graph"
@@ -28,16 +31,65 @@ var (
 
 // BuildTestBackend and sets backends.DefaultConfig to "xla:cpu" -- it can be overwritten by GOMLX_BACKEND environment variable.
 func BuildTestBackend() backends.Backend {
-	backends.DefaultConfig = "xla:cpu"
+	backends.DefaultConfig = officialTestBackendNames[0]
 	backendOnce.Do(func() {
 		err := xla.AutoInstall()
 		if err != nil {
 			klog.Fatalf("Failed to auto-install XLA PJRT: %+v", err)
 		}
-		cachedBackend = backends.MustNew()
-		fmt.Printf("Backend: %s\n", cachedBackend.Description())
+		for i, backendName := range officialTestBackendNames {
+			backend, err := backends.NewWithConfig(backendName)
+			if err != nil {
+				if i == 0 {
+					klog.Fatalf("Failed to create backend %q: %+v", backendName, err)
+				}
+				klog.Errorf("Failed to create backend %q: %+v", backendName, err)
+				continue
+			}
+			officialTestBackends[backendName] = backend
+		}
 	})
-	return cachedBackend
+	return officialTestBackends[backends.DefaultConfig]
+}
+
+var (
+	officialTestBackendNames []string = []string{
+		"xla:cpu",
+		"go",
+	}
+	officialTestBackends = make(map[string]backends.Backend)
+)
+
+func init() {
+	if selectedBackendName := os.Getenv(backends.ConfigEnvVar); selectedBackendName != "" {
+		officialTestBackendNames = []string{selectedBackendName}
+		return
+	}
+
+	// Include CUDA PJRT test if available.
+	if installer.HasNvidiaGPU() {
+		officialTestBackendNames = append(officialTestBackendNames, "xla:cuda")
+	}
+}
+
+// TestOfficialBackends iterates over list of backends and calls testFn for each of them.
+// If GOMLX_BACKEND environment variable is set, it will only iterate over the one set.
+// If GOMLX_BACKEND is not set, it will iterate over all official backends, except those in excludeBackends.
+// (for tests known not to work on those backends)
+func TestOfficialBackends(t *testing.T, testFn func(t *testing.T, backend backends.Backend), excludeBackends ...string) {
+	BuildTestBackend()
+	for backendName, backend := range officialTestBackends {
+		if slices.Contains(excludeBackends, backendName) {
+			continue
+		}
+		if backend == nil {
+			// This happens if the backend already failed to initialize, no need to report it more than once.
+			continue
+		}
+		t.Run(backendName, func(t *testing.T) {
+			testFn(t, backend)
+		})
+	}
 }
 
 // RunTestGraphFn tests a graph building function graphFn by executing it and comparing
@@ -46,8 +98,16 @@ func BuildTestBackend() backends.Backend {
 // delta is the margin of value on the difference of output and want values that are acceptable.
 // Values of delta <= 0 means only exact equality is accepted.
 func RunTestGraphFn(t *testing.T, testName string, graphFn TestGraphFn, want []any, delta float64) {
+	RunTestGraphFnWithBackend(t, testName, BuildTestBackend(), graphFn, want, delta)
+}
+
+// RunTestGraphFnWithBackend tests a graph building function graphFn by executing it and comparing
+// its output(s) to the values in want, reporting back any errors in t.
+//
+// delta is the margin of value on the difference of output and want values that are acceptable.
+// Values of delta <= 0 means only exact equality is accepted.
+func RunTestGraphFnWithBackend(t *testing.T, testName string, backend backends.Backend, graphFn TestGraphFn, want []any, delta float64) {
 	t.Run(testName, func(t *testing.T) {
-		backend := BuildTestBackend()
 		wantTensors := xslices.Map(want, func(value any) *tensors.Tensor {
 			if s, ok := value.(shapes.Shape); ok {
 				return tensors.FromShape(s)